Rename Generator to TransformIterator
This iterator serves as a wrapper for lightweight iterators (objects that have next, hasNext, and reset methods). The lightweight iterators gain all of the array-like methods and convenience methods. It's like a cheap sub-class.

Transform iterators also allow a function to be applied to their members potentially serving as a form of laziness. Although, the equivalent laziness can be recreated using "take" and then Array.map.